+++ examples/multiFileProject/multiFileProject.cpp
@@ -1,0 +1,13 @@
+/****************************************************************************************************************************
+  multiFileProject.cpp
+  
+  For ESP32, ESP32_S2, ESP32_S3, ESP32_C3 boards with ESP32 core v2.0.2+
+  Written by Khoi Hoang
+
+  Built by Khoi Hoang https://github.com/khoih-prog/ESP32TimerInterrupt
+  Licensed under MIT license
+*****************************************************************************************************************************/
+
+// To demo how to include files in multi-file Projects
+
+#include "multiFileProject.h"